perf: keep setup-guides off the training-plan client bundle via server shell
